refactor: папка hooks вынесена на один уровень с компонентами
